fixes compile with verbosity (and nightly tests)
The current solution is not ideal as the verbosity header are located
in the micoscale. The header should be moved to a util folder that both
macro and micro link against, but this gets into creating header only
targets in cmake which I wanted to avoid for now. This will need to be
revisited with #103