clean memory leaks & move sparskit to own structure (#117)
* move sparskit arrays to own structure

close #113

* macroscale leaks

* remove the mesh field from the FiberReactions struct

The mesh should not be located in this struct because the fiber
reactions can also refer to mesh copies. E.g. the user should have no
expectation that the mesh located in the FiberReactions structure is the
correct mesh, so it should not be located there.

* fix microscale leaks

* address some of @tobinw's comments

* fix test cases to work with new structs

The test cases needed some slight modification to work with the new
structures. Also, there was a but in the csr test case caused by ending
the incorrect mesh iterator. This caused an error upon destructing the
mesh.

* rename FiberRVEAnalysisVecs to LinearStrucs

* Fix clang format file and bump patch number.