Material Stiffness Derivs/Neohookean at microscale
This commit modifies the multiscale coupling to use the material
stiffness from the microscale to compute the derivatives at the
macroscale. In order to test this functionality, we implement a
neohookean model at the microscale (closes #56). Due to the more general
class structure which was implemented this commit closes #134.

In theory this commit should allow using multiple types of RVEs at the
microscale, however this functionality has not been tested yet (closes #103).

This commit closes #132 since we no longer are using the dx_fn_dx_rve
derivative terms.