Make kokkos a required library
As we continue to refactor the code, we plan to make all of our
underlying data structures use kokkos views or kokkos dual views.
This commit removes all of the ifdefs that enable compiling without
Kokkos, because this will not be an option moving forward. Additionally,
I transition to use the Kokkos::ScopeGuard functionality