Restructure the batched analysis code
Here we somewhat restructure the batched analysis code. This seems
to strongly pessimize any sort of vectorization that the compiler
was doing on the serial loops and we went from 2x pure serial wot 3x
pure serial with the Kokkos serial backend. This will need some work
once the cuda team backend is functioning.