Add UpdatedLagrangianMaterialIntegrator
Based on bug I found with the coordinates looking at issue #31 I spent
some time looking at the MaterialIntegrator code (Neohookean and
ULMultiscale). It's clear that the repeated code needed to be abstracted
out. This new approach allows the material definitions to be defined
independently of the integration proceedures needed to construct the
updated Lagrangian finite elements. This *should* improve testability.