Test UpdatedLagrangianMaterialIntegrator and support
This commit adds a bunch of tests including neohookean material tests
that close #33. These tests are made possible by the updated design
using UpdatedLagrangianMaterialIntegrator. So far, the
UpdatedLagrangianMaterialIntegrator has not been used in the code to
replace the ULMultiscaleIntegrator and NeoHookeanIntegrator. However,
now that these tests are passing this is the sensible next step.

This design change nicely separates the concerns of the FEM stuff needed
for the updated lagrangian formulation from the material definitions.